+++ .idea/dictionaries/josep.xml
@@ -1,0 +1,8 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="josep">
+    <words>
+      <w>bonuspoints</w>
+      <w>bonuswords</w>
+    </words>
+  </dictionary>
+</component>
